requirements.txt
streamlit>=1.45.0
google-genai>=1.14.0
openai>=1.30.0
supabase>=2.15.0
Pillow>=11.0.0
PyYAML>=6.0.2
python-dotenv>=1.1.0
pytest>=8.3.0
